32b

+++ models/llm/cuda_backend.py
@@ -15,7 +15,7 @@ class CUDABackend(LLMBackend):
         self.tokenizer = AutoTokenizer.from_pretrained(model_name)
         self.model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=torch.float16,
+            dtype=torch.bfloat16,
             device_map="auto",
         )
